Auto Improvements

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Motors.java
@@ -12,8 +12,8 @@ public final class Motors {
     public static final int ANDYMARK_20_CPR = 560;
     public static final int ANDYMARK_40_CPR = 1120;
     public static final int ANDYMARK_60_CPR = 1680;
-    public static final int ANDYMARK_3_7_CPR = (int) 44.4;
+    public static final double ANDYMARK_3_7_CPR = 44.4;
     public static final int TETRIX_CPR = 1440;
     public static final int REV_HD_HEX_CPR = 2240;
-    public static final int MODERNROBOTICS_MATRIX_12V_CPR = (int) 1478.4;
+    public static final double MODERNROBOTICS_MATRIX_12V_CPR = 1478.4;
 }
